shooter updates
